add: funding settings
wip: #1341

add: new page on form wizard, changed form name, yup validation, etc

wip: split page with qf settings

add: form registers
