+++ packages/round-manager/src/features/round/__tests__/CreateRoundPage.test.tsx
@@ -7,6 +7,7 @@ import { RoundApplicationForm } from "../RoundApplicationForm";
 import { useWallet } from "../../common/Auth";
 import * as FormWizardImport from "../../common/FormWizard";
 import { fireEvent, screen } from "@testing-library/react";
+import QuadraticFundingForm from "../QuadraticFundingFrom";
 
 jest.mock("../../common/Navbar");
 jest.mock("../../common/Auth");
@@ -42,6 +43,7 @@ describe("<CreateRoundPage />", () => {
     expect(firstCallArgument).toMatchObject({
       steps: [
         RoundDetailForm,
+        QuadraticFundingForm,
         ApplicationEligibilityForm,
         RoundApplicationForm,
       ],
@@ -71,6 +73,7 @@ describe("<CreateRoundPage />", () => {
     expect(firstCallArgument).toMatchObject({
       steps: [
         RoundDetailForm,
+        QuadraticFundingForm,
         ApplicationEligibilityForm,
         RoundApplicationForm,
       ],
